new task editor
