+++ web/src/app/components/tasks-editor/tasks-editor.component.html
@@ -14,16 +14,28 @@
   limitations under the License.
 -->
 
-<div id="task-details">
-  <h1>
-    <div *ngIf="label">Job: {{ label }}</div>
-    What data should be collected?
-  </h1>
+<div id="tasks-added"
+  *ngIf="formArray"
+  cdkDropList
+  (cdkDropListDropped)="drop($event)"
+>
+  <task-form
+    *ngFor="let formGroup of formArray.controls; let i = index;"
+    [formGroup]="formGroup"
+    [index]="i"
+    (delete)="onTaskDelete(i)"
+    (duplicate)="onTaskDuplicate(i)"
+  >
+  </task-form>
+</div>
 
-  <h2>
-    Data collectors will be prompted to complete the tasks you define here for
-    each data collection site.
-  </h2>
+<div id="tasks-help-text">Select task to add</div>
 
-  <tasks-editor #tasksEditor (onChange)="onChange($event)"></tasks-editor>
+<div id="add-task-buttons-list">
+  <add-task-button
+    *ngFor="let taskGroup of addableTaskGroups"
+    [taskGroup]="taskGroup"
+    (click)="onTaskAdd(taskGroup)"
+  >
+  </add-task-button>
 </div>
