+++ web/src/app/components/tasks-editor/tasks-editor.component.scss
@@ -13,3 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+#tasks-help-text {
+  padding-top: 3ch;
+  padding-bottom: 2ch;
+  text-align: center;
+  color: #727970;
+}
+
+#add-task-buttons-list {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 8px;
+}
