+++ web/src/app/pages/edit-survey/edit-job/edit-job.component.ts
@@ -14,12 +14,7 @@
  * limitations under the License.
  */
 
-import {CdkDragDrop} from '@angular/cdk/drag-drop';
 import {Task} from 'app/models/task/task.model';
-import {
-  TaskGroup,
-  taskGroupToTypes,
-} from 'app/pages/create-survey/task-details/task-details.component';
 import {DialogService} from 'app/services/dialog/dialog.service';
 import {NavigationService} from 'app/services/navigation/navigation.service';
 import {SurveyService} from 'app/services/survey/survey.service';
@@ -41,15 +36,7 @@ export class EditJobComponent {
   surveyId?: string;
   jobId?: string;
   section: 'tasks' | 'lois' = 'tasks';
-
   tasks?: List<Task>;
-  addableTaskGroups: Array<TaskGroup> = [
-    TaskGroup.QUESTION,
-    TaskGroup.PHOTO,
-    TaskGroup.DROP_PIN,
-    TaskGroup.DRAW_AREA,
-    TaskGroup.CAPTURE_LOCATION,
-  ];
 
   @ViewChild('loiSelection')
   loiSelection?: LoiSelectionComponent;
@@ -103,87 +90,6 @@ export class EditJobComponent {
 
   onChangeSection(section: 'tasks' | 'lois') {
     this.section = section;
-  }
-
-  onAddTask(group: TaskGroup) {
-    const types = taskGroupToTypes.get(group);
-
-    const type = types?.first();
-
-    if (type && this.tasks) {
-      const task = this.taskService.createTask(
-        type,
-        '',
-        false,
-        this.tasks.size
-      );
-
-      this.tasks = this.tasks.push(task);
-    }
-
-    // No need to call addOrUpdateTasks because ngOnChange emits an update once
-    // the task-input component is created.
-  }
-
-  onUpdateTask(event: Task, index: number) {
-    if (!this.tasks) {
-      throw Error('tasks list is is empty');
-    }
-
-    this.tasks = this.tasks.set(index, event);
-    this.taskService.addOrUpdateTasks(this.surveyId!, this.jobId!, this.tasks);
-  }
-
-  onDeleteTask(index: number) {
-    this.dialogService
-      .openConfirmationDialog(
-        'Warning',
-        'Are you sure you wish to delete this question? Any associated data ' +
-          'will be lost. This cannot be undone.'
-      )
-      .afterClosed()
-      .subscribe(dialogResult => {
-        if (dialogResult) {
-          this.tasks = this.tasks!.splice(index, 1);
-        }
-      });
-  }
-
-  drop(event: CdkDragDrop<string[]>): void {
-    const {previousIndex, currentIndex} = event;
-    this.tasks = this.tasks!.update(
-      previousIndex,
-      task => task?.copyWith({index: currentIndex}) as Task
-    )
-      .update(
-        currentIndex,
-        task => task?.copyWith({index: previousIndex}) as Task
-      )
-      .sortBy(task => task.index);
-  }
-
-  onDuplicateTask(index: number) {
-    this.dialogService
-      .openConfirmationDialog(
-        'Duplicate task',
-        'Are you sure you wish to duplicate this task?'
-      )
-      .afterClosed()
-      .subscribe(dialogResult => {
-        if (dialogResult) {
-          const taskToDuplicate = this.tasks!.get(index);
-          if (taskToDuplicate) {
-            const task = this.taskService.createTask(
-              taskToDuplicate?.type,
-              taskToDuplicate?.label,
-              taskToDuplicate?.required,
-              this.tasks!.size,
-              taskToDuplicate?.multipleChoice
-            );
-            this.tasks = this.tasks!.push(task);
-          }
-        }
-      });
   }
 
   ngOnDestroy() {
