+++ web/src/app/pages/edit-survey/edit-job/edit-job.module.ts
@@ -17,23 +17,19 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {EditJobComponent} from 'app/pages/edit-survey/edit-job/edit-job.component';
-import {TaskInputModule} from 'app/pages/create-survey/task-details/task-input/task-input.module';
-import {AddTaskButtonModule} from 'app/pages/create-survey/task-details/add-task-button/add-task-button.module';
-import {DragDropModule} from '@angular/cdk/drag-drop';
 import {LoiSelectionModule} from 'app/pages/create-survey/loi-selection/loi-selection.module';
 import {MatButtonModule} from '@angular/material/button';
 import {MatIconModule} from '@angular/material/icon';
+import {TasksEditorModule} from 'app/components/tasks-editor/tasks-editor.module';
 
 @NgModule({
   declarations: [EditJobComponent],
   imports: [
     CommonModule,
-    TaskInputModule,
-    AddTaskButtonModule,
-    DragDropModule,
     LoiSelectionModule,
     MatButtonModule,
     MatIconModule,
+    TasksEditorModule,
   ],
   exports: [EditJobComponent],
 })
